// src/app/models/data/linFrame.ts
export class LinFrame
{
    SessionID: string;
    PCKNO: number;
    FNO : number;
    PID: string;
    PIDDec?: string;
    PIDName?: string;
    FDATA0: string;
    FDATA0Name?: string;
    FDATA1: string;
    FDATA1Name?: string;
    FDATA2: string;
    FDATA2Name?: string;
    FDATA3: string;
    FDATA3Name?: string;
    FDATA4: string;
    FDATA4Name?: string;
    FDATA5: string;
    FDATA5Name?: string;
    FDATA6: string;
    FDATA6Name?: string;
    FDATA7: string;
    FDATA7Name?: string;
}